+++ acceptance-tests/src/test/java/tech/pegasys/ethsigner/tests/signing/ValueTransferAcceptanceTest.java
@@ -13,6 +13,7 @@
 package tech.pegasys.ethsigner.tests.signing;
 
 import static io.netty.handler.codec.http.HttpResponseStatus.BAD_REQUEST;
+import static io.netty.handler.codec.http.HttpResponseStatus.OK;
 import static org.assertj.core.api.Assertions.assertThat;
 import static tech.pegasys.ethsigner.core.jsonrpc.response.JsonRpcError.SIGNING_FROM_IS_NOT_AN_UNLOCKED_ACCOUNT;
 import static tech.pegasys.ethsigner.core.jsonrpc.response.JsonRpcError.TRANSACTION_UPFRONT_COST_EXCEEDS_BALANCE;
@@ -75,7 +76,7 @@ public class ValueTransferAcceptanceTest extends AcceptanceTestBase {
 
     final SignerResponse<JsonRpcErrorResponse> signerResponse =
         ethSigner().transactions().submitExceptional(transaction);
-    assertThat(signerResponse.status()).isEqualTo(BAD_REQUEST);
+    assertThat(signerResponse.status()).isEqualTo(OK);
     assertThat(signerResponse.jsonRpc().getError())
         .isEqualTo(TRANSACTION_UPFRONT_COST_EXCEEDS_BALANCE);