+++ ethsigner/core/src/main/java/tech/pegasys/ethsigner/core/requesthandler/VertxRequestTransmitter.java
@@ -17,7 +17,10 @@ import tech.pegasys.ethsigner.core.requesthandler.sendtransaction.DownstreamPath
 import java.nio.charset.StandardCharsets;
 import java.time.Duration;
 import java.util.Map.Entry;
+import java.util.Objects;
 import java.util.concurrent.atomic.AtomicBoolean;
+import java.util.stream.Collectors;
+import java.util.stream.StreamSupport;
 
 import io.vertx.core.Vertx;
 import io.vertx.core.http.HttpClient;
@@ -57,7 +60,15 @@ public class VertxRequestTransmitter implements RequestTransmitter {
       final Iterable<Entry<String, String>> headers,
       final String path,
       final String body) {
-    LOG.debug("Sending request {} to {}", body, path);
+    LOG.debug(
+        "Sending headers {} and request {} to {} ",
+        () ->
+            StreamSupport.stream(headers.spliterator(), false)
+                .map(Objects::toString)
+                .collect(Collectors.joining(", ")),
+        () -> body,
+        () -> path);
+
     final String fullPath = downstreamPathCalculator.calculateDownstreamPath(path);
     final HttpClientRequest request =
         downStreamConnection.request(method, fullPath, this::handleResponse);